requirements.txt
flask
transformers
torch
evaluate
nltk
pymupdf
pytesseract
Pillow
sentencepiece
absl-py
rouge-score
